feat: 添加五彩纸屑效果

+++ app/screens/DataMigratorScreen/helpers/clearOldData.ts
@@ -1,7 +1,11 @@
-import { unlink } from 'react-native-fs';
+import { exists, unlink } from 'react-native-fs';
 import { DATA_PATH, OLD_DB_PATH } from '../constants';
 
-export function clearOldData() {
-  unlink(DATA_PATH);
-  unlink(OLD_DB_PATH);
+export async function clearOldData() {
+  if (await exists(DATA_PATH)) {
+    await unlink(DATA_PATH);
+  }
+  if (await exists(OLD_DB_PATH)) {
+    await unlink(OLD_DB_PATH);
+  }
 }
